+++ alpha-test-build-dist.sh
@@ -1,15 +1,7 @@
 #!/bin/bash
 # Run CI test & build for a local alpha release
 set -euxo pipefail
-CODEX_BASE_VERSION=$(./docker-version-codex-base.sh)
-export CODEX_BASE_VERSION
-CODEX_BUILDER_BASE_VERSION=$(./docker-version-codex-builder-base.sh)
-export CODEX_BUILDER_BASE_VERSION
-CODEX_DIST_BUILDER_VERSION=$(./docker-version-codex-dist-builder.sh)
-export CODEX_DIST_BUILDER_VERSION
-CODEX_BUILDER_FINAL_VERSION=$(./docker-version-codex-builder-final.sh)
-export CODEX_BUILDER_FINAL_VERSION
-./docker-build-codex-dist-builder.sh
+./docker/docker-build-image.sh codex-dist-builder
 ./docker/docker-compose-exit.sh codex-save-cache
 ./docker/docker-compose-exit.sh codex-frontend-lint
 ./docker/docker-compose-exit.sh codex-frontend-test