+++ docker/docker-version-codex-builder-final.sh
@@ -1,0 +1,17 @@
+#!/bin/bash
+# Compute the version tag for codex-builder-final
+set -euo pipefail
+
+# shellcheck disable=SC1091
+source .env
+EXTRA_MD5S=(
+    "$CODEX_BUILDER_BASE_VERSION  codex-builder-base-version"
+    "$PKG_VERSION  codex-package-version")
+
+# shellcheck disable=SC2046
+DEPS=(
+    "$0"
+    builder-final.Dockerfile
+)
+
+source ./docker/docker-version-checksum.sh