+++ fix-lint.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -euo pipefail
-./fix-lint-frontend.sh
+./frontend/fix-lint.sh
 ./fix-lint-backend.sh